Added argument for fpga_top.v path

+++ core_gen.py
@@ -1,14 +1,26 @@
+import sys
 import os
 
+if len(sys.argv) > 1:
+    top_path = sys.argv[1]
+    print("\033[1;32mfpga_top\033[m location supplied as a script argument: \033[1;32m", top_path, "\033[m")
+else:
+    top_path = input('Input location of \033[1;32mfpga_top.v\033[m, type \"local\" for local directory, or type \"default\" for default.\n> ')
+    if top_path == "default" or top_path == "":
+        top_path = 'fpga_top.v'
+
 core_path = "fpga_core.v"
+
 if(os.path.exists(core_path)):
-    os.remove(core_path)
+    rem = input('\nExisting fpga_core.v found on core path. Do you want to overwrite it? Type yes or no\n >')
+    if(rem == "yes"):
+        os.remove(core_path)
+    else:
+        print("Aborting...")
+        sys.exit()
 
 print("This script takes fpga_top.v as input and generates a valid fpga_core.v.")
-top_path = input('Input location of \033[1;32mfpga_top.v\033[m, type \"local\" for local directory, or type \"default\" for default.\n> ')
 
-if top_path == "default" or top_path == "":
-    top_path = 'fpga_top.v'
 
 with open(top_path, 'r', encoding="utf-8") as top:
     top_data = top.read();
@@ -37,7 +49,3 @@ for top_word in find_all(top_data, "grid_io_"):
     core_file = open(core_path, "a")
     core_file.write(top_data[start_index:top_word])
     start_index = end_index
-
-
-
-
